pyproject.toml
[project]
name = "kr-rest-days"
version = "0.1.0"
description = "한국 공휴일 정보 조회 서비스"
requires-python = ">=3.11"
authors = [
    { name = "Gemini CLI User" }
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "loguru>=0.7.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
